FIX: Minor tweaks
