+++ doc/release/0.19.0-notes.rst
@@ -22,6 +22,11 @@ This release requires Python 2.7 or 3.4-3.5 and NumPy 1.7.1 or greater.
 New features
 ============
 
+`scipy.signal` improvements
+---------------------------
+
+The function `scipy.signal.sosfreqz` was added to compute the frequency
+response from second-order sections.
 
 
 Deprecated features
